Extract helpers + prepare/bump next version
