+++ packages/xchain-thorchain/CHANGELOG.md
@@ -1,3 +1,13 @@
+# v.0.19.0 (2021-09-10)
+
+## Breaking change
+
+- Extract `buildDepositTx` from `Client` into `utils`
+
+## Add
+
+- Add `getBalance` to `utils`
+
 # v.0.18.0 (2021-09-08)
 
 ## Add
